agents/perplexity-intel-agent/requirements.txt
anthropic
tavily-python
python-dotenv
